refact: clickhouse code
