+++ backend/pkg/dbmanager/postgresql_utils.go
@@ -102,3 +102,51 @@ func extractTableName(stmt string) string {
 
 	return tableName
 }
+
+// Update the function signature to include indexes
+func convertTablesToSchemaFormat(tables map[string]PostgresTable, indexes map[string][]PostgresIndex) map[string]TableSchema {
+	result := make(map[string]TableSchema)
+	for name, table := range tables {
+		schema := TableSchema{
+			Name:        name,
+			Columns:     make(map[string]ColumnInfo),
+			Indexes:     make(map[string]IndexInfo),
+			ForeignKeys: make(map[string]ForeignKey),
+			Constraints: make(map[string]ConstraintInfo),
+			RowCount:    table.RowCount,
+		}
+
+		// Convert columns
+		for colName, col := range table.Columns {
+			schema.Columns[colName] = col.toColumnInfo()
+		}
+
+		// Convert indexes
+		if tableIndexes, ok := indexes[name]; ok {
+			for _, idx := range tableIndexes {
+				schema.Indexes[idx.Name] = IndexInfo{
+					Name:     idx.Name,
+					Columns:  idx.Columns,
+					IsUnique: idx.IsUnique,
+				}
+			}
+		}
+
+		// Convert foreign keys
+		if table.ForeignKeys != nil {
+			for fkName, fk := range table.ForeignKeys {
+				schema.ForeignKeys[fkName] = ForeignKey{
+					Name:       fk.Name,
+					ColumnName: fk.Column,
+					RefTable:   fk.RefTable,
+					RefColumn:  fk.RefColumn,
+					OnDelete:   fk.OnDelete,
+					OnUpdate:   fk.OnUpdate,
+				}
+			}
+		}
+
+		result[name] = schema
+	}
+	return result
+}
